fix misconception with how vertex-array (VAO) work
I thought the glBindBuffer(...) were part of the vertex array: this is
incorrect!

Vertex Array Objects (VAO) only store what attributes are enabled/disabled and
to what data they are bound
to (glEnableVertexAttribArray/glVertexAttribPointer)

The glBindBuffer we perform while the VAO is bound is only there to
allow the glVertexAttribPointer to know which buffer to work with. They
do not carry on.

This also mean I was incorrect to bind the GL_ELEMENT_ARRAY_BUFFER in
that scope: it should be done when drawing (calling glDrawElements)